pyproject.toml
[project.urls]
homepage = "https://github.com/han-xudong/magiclaw"
issues = "https://github.com/han-xudong/magiclaw/issues"

[project]
name = "magiclaw"
version = "0.1.0"
description = "MagiClaw: Universal Action Embodiment"
authors = [
    {name = "Xudong Han", email = "12231112@mail.sustech.edu.cn"},
    {name = "Tianyu Wu", email = "12431162@mail.sustech.edu.cn"},
    {name = "Sheng Ge", email = "12132253@mail.sustech.edu.cn"},
    {name = "Yang Xiao", email = "11811019@mail.sustech.edu.cn"},
    {name = "Yujian Dong", email = "12232321@mail.sustech.edu.cn"},
    {name = "Ronghan Xu", email = "12232320@mail.sustech.edu.cn"},
    {name = "Junlin Wang", email = "12112921@mail.sustech.edu.cn"},
    {name = "Fang Wan", email = "wanf@sustech.edu.cn"},
    {name = "Chaoyang Song", email = "songcy@ieee.org"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["robotics", "deep learning", "manipulation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Robotics",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "jupyter>=1.1.1",
    "numpy>=2.2.4",
    "onnx>=1.17.0",
    "onnxruntime>=1.21.0",
    "opencv-contrib-python>=4.11.0.86",
    "opencv-python>=4.11.0.86",
    "protobuf>=6.30.1",
    "pylkmotor>=1.0.4",
    "python-can>=4.5.0",
    "pyyaml>=6.0.2",
    "pyzmq>=26.2.0",
    "scapy>=2.6.1",
    "scipy>=1.15.2",
]

[tool.poetry]
requires-poetry = ">=2.1"

[tool.ruff]
line-length = 110
target-version = "py310"
exclude = [
    "tests/data",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "N", "B", "C4", "SIM"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"